Move registerNodeToBP before RPC server listen
