+++ cmd/cql-observer/observation_test.go
@@ -233,11 +233,25 @@ func TestFullProcess(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
 
 	Convey("test full process", t, func() {
+		var err error
 		startNodes()
 		defer stopNodes()
+		ctx, cancel := context.WithTimeout(context.Background(), time.Second*30)
+		defer cancel()
+		err = utils.WaitToConnect(ctx, "127.0.0.1", []int{
+			4120,
+			4121,
+			4122,
+			4144,
+			4145,
+			4146,
+		}, time.Millisecond*200)
+		if err != nil {
+			log.Fatalf("wait for port ready timeout: %v", err)
+		}
+
 		time.Sleep(10 * time.Second)
 
-		var err error
 		err = client.Init(FJ(testWorkingDir, "./observation/node_c/config.yaml"), []byte(""))
 		So(err, ShouldBeNil)
